requirements.txt
asgiref==3.5.0
Django==4.0.3
django-appconf==1.0.5
django-filter==21.1
django-imagekit==4.1.0
djangorestframework==3.13.1
pilkit==2.0
Pillow==9.0.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
